requirements.txt
requests
mysql-connector-python
PyYAML
